msg: improve message handling
This is a huge commit and its purpose is to split code and text. 'text' means
the text that the user will get in case of a test failure. This way, code and
text do not mix and can work in parallel. For those tests pending implementation,
a TODO mark is given.

Signed-off-by: Leonardo Sandoval <leonardo.sandoval.gonzalez@linux.intel.com>