summary: remove the summary & doc/fixes dependency
Instead of having a separate file containing 'proposed fixes', these strings
will be contained inside the test case itself. Possible contributors to the
suite will find it easier to edit just one file (the test case) instead of the
test case AND the fixes file. Also, this keeps the fixes string more accurate
if these are closer to the code.

Signed-off-by: Leonardo Sandoval <leonardo.sandoval.gonzalez@linux.intel.com>