tests: adjust tests to support new summary format
The new patchtest summary format requires us to format the results (in
particular the failures) in a certain way. Most tests are expected to
return a description of the problem ("issue"), an optional suggested
way to fix the problem ("fix") and if applicable the commit which the
failure applies to.

As part of this, the fail() function has been changed to accept
issue, fix and commit parameters since these are the most common. The
fix was previously a decorator, but if we can raise multiple failures
per test, it is logical that those failures each could have a slightly
different suggested fix.

Additionally, the fail() function now assembles the results into a dict
since the jinja2 template is now in charge of formatting it, so the
order is no longer important (however the "data" free-form list is still
kept as a list of tuples since the order of that may still be
significant).

I have added issue parameters to all of the fail() calls and along the
way I have also tweaked a number of the suggested fixes. As a general
rule I believe we should be expressing additional information as part of
the issue / fix text rather than supplying it in field form via the data
parameter for better human readability.

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>