Improve API for accessing commit information
Rather than a set of same-indexed lists, use a single list of namedtuple
objects to get access to the shortlog, commit message and payload for
each commit/patch in the patchset. This tidies up the code a little.

(I was initially going to name this "patches" but we already have a
"patchset" object which would seem like a bit of a name clash. Perhaps
the two can be brought together at some point in the future though.)

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>